requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Deprecated==1.2.12
Django==3.2.5
django-cors-headers==3.8.0
django-cors-middleware==1.5.0
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
exponent-server-sdk==2.0.0
google-auth==2.3.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.4
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
